backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.4
python-dotenv==1.0.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
stripe==11.0.0
requests==2.32.3
